Show/Close properly modal - change target lang
Add functionality of changing input and output language.
Clicking each language opens a modal with list of languages.

Create modal that appears in the middle. List of languages is
created when the button is clicked - in JS from object holding
each language and its language code.

<> Still need to style those displayed languages and find a way
to set language to the proper target button.