Implement search language. Match 1st letter
When Modal is opened, first needed to focus on the modal container
Then listening on keypress event.

With given key, filtered out all languages starting with that letter.
First: lookup if any of those languages is focused, then focus on
the next one.

After: if none language from this group is focused, then focus
on the first one of that group