Add div to body. Translation above selection
For now this is the best solution I found about displaying translation
above the selected content.
Tried before by wrapping selection in `span` and using ::before to
display small tooltip above, but it wasn't working in many cases
(hiding under other boxes).

This creates div and appends it to body with absolute position.
Div placed above selected content (calculated from selection and
distance from the top of the page);

From this point, needs to add buttons to the div, and improve styles