Translate selected txt to English by default
When extension is installed it is not possible to translate selected
text, because Event script pulls output language from the memory and
language is not set before user chooses it manually.

Changed it to translate by default to English, so it should force user
to open popup and change there manually.